requirements.txt
python-telegram-bot
pandas
requests
pandas-ta
apscheduler
